Pull object SchoolYearGraduationSerializers into module that is shared
